+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## Unreleased
+## 22.8.2
 
 New error codes:
 * Y047: Detect unused `TypeAlias` declarations.